Remove hardcoded constants from window services
