+++ src/helm/benchmark/window_services/image_generation/test_clip_window_service.py
@@ -1,6 +1,7 @@
 import shutil
 import tempfile
 
+from helm.benchmark.config_registry import register_builtin_configs_from_helm_package
 from helm.benchmark.window_services.tokenizer_service import TokenizerService
 from helm.benchmark.window_services.test_utils import get_tokenizer_service
 from helm.benchmark.window_services.window_service_factory import WindowServiceFactory
@@ -8,6 +9,7 @@ from helm.benchmark.window_services.window_service_factory import WindowServiceF
 
 class TestCLIPWindowService:
     def setup_method(self):
+        register_builtin_configs_from_helm_package()
         self.path: str = tempfile.mkdtemp()
         service: TokenizerService = get_tokenizer_service(self.path)
         self.window_service = WindowServiceFactory.get_window_service("huggingface/dreamlike-photoreal-v2-0", service)
