+++ src/helm/benchmark/window_services/image_generation/test_openai_dalle_window_service.py
@@ -1,6 +1,7 @@
 import shutil
 import tempfile
 
+from helm.benchmark.config_registry import register_builtin_configs_from_helm_package
 from helm.benchmark.window_services.tokenizer_service import TokenizerService
 from helm.proxy.clients.image_generation.dalle2_client import DALLE2Client
 from helm.benchmark.window_services.test_utils import get_tokenizer_service, TEST_PROMPT
@@ -9,6 +10,7 @@ from helm.benchmark.window_services.window_service_factory import WindowServiceF
 
 class TestOpenAIDALLEWindowService:
     def setup_method(self):
+        register_builtin_configs_from_helm_package()
         self.path: str = tempfile.mkdtemp()
         service: TokenizerService = get_tokenizer_service(self.path)
         self.window_service = WindowServiceFactory.get_window_service("openai/dall-e-2", service)
