+++ src/helm/benchmark/window_services/test_cohere_window_service.py
@@ -6,6 +6,7 @@ from typing import List
 
 from sqlitedict import SqliteDict
 
+from helm.benchmark.config_registry import register_builtin_configs_from_helm_package
 from helm.common.general import ensure_directory_exists
 from .test_cohere_window_service_utils import REQUESTS_TO_RESPONSES, TEST_PROMPT, TOKENIZED_PROMPT
 from .tokenizer_service import TokenizerService
@@ -16,6 +17,7 @@ from .test_utils import get_tokenizer_service
 class TestCohereWindowService:
     @classmethod
     def setup_class(cls):
+        register_builtin_configs_from_helm_package()
         cls.path: str = tempfile.mkdtemp()
         cache_path: str = os.path.join(cls.path, "cache")
         ensure_directory_exists(cache_path)
