+++ src/helm/benchmark/window_services/test_opt_window_service.py
@@ -1,6 +1,7 @@
 import shutil
 import tempfile
 
+from helm.benchmark.config_registry import register_builtin_configs_from_helm_package
 from .test_utils import get_tokenizer_service, TEST_PROMPT
 from .tokenizer_service import TokenizerService
 from .window_service_factory import WindowServiceFactory
@@ -8,6 +9,7 @@ from .window_service_factory import WindowServiceFactory
 
 class TestOPTWindowService:
     def setup_method(self):
+        register_builtin_configs_from_helm_package()
         self.path: str = tempfile.mkdtemp()
         service: TokenizerService = get_tokenizer_service(self.path)
         self.window_service = WindowServiceFactory.get_window_service("together/opt-175b", service)
