+++ src/helm/benchmark/window_services/test_palmyra_window_service.py
@@ -1,6 +1,7 @@
 from tempfile import TemporaryDirectory
 from typing import List
 
+from helm.benchmark.config_registry import register_builtin_configs_from_helm_package
 from .tokenizer_service import TokenizerService
 from .window_service_factory import WindowServiceFactory
 from .test_utils import get_tokenizer_service, TEST_PROMPT
@@ -116,6 +117,7 @@ class TestPalmyraWindowService:
     ]
 
     def setup_method(self):
+        register_builtin_configs_from_helm_package()
         self.temporary_directory = TemporaryDirectory()
         service: TokenizerService = get_tokenizer_service(self.temporary_directory.name)
         self.window_service = WindowServiceFactory.get_window_service("writer/palmyra-large", service)
